brainscore_vision/models/hmax_v3_adj/requirements.txt
torch
torchvision
urllib3
numpy
pytest
pyyaml
huggingface_hub
safetensors
requests
timm
